Moved trade trait impl to trades from utilities
