+++ src/markets.rs
@@ -2,7 +2,8 @@ use crate::candles::{select_first_01d_candle, ResearchCandle};
 use crate::eldorado::ElDorado;
 use crate::exchanges::{client::RestClient, error::RestError, select_exchanges, ExchangeName};
 use crate::inquisidor::Inquisidor;
-use crate::utilities::{get_input, TimeFrame, Trade};
+use crate::trades::Trade;
+use crate::utilities::{get_input, TimeFrame};
 use chrono::{DateTime, Duration, Utc};
 use core::cmp::Reverse;
 use rust_decimal::Decimal;
